requirements.txt
tweepy
pandas
google-cloud-bigquery
pyarrow
